Updated test_category_sikraken.sh To Add A Specified Timeout
test_category_sikraken.sh used to have no timeout when calling to ECLiPSe. ECLiPSe may not terminate itself despite being given a time budget so this is just an extra precaution. The current timeout duration is set to 1 second to show it does work and can be changed as required. The error messages however may need to be looked into as parallel execution may fail but no message will show up in the elif statement.